+++ src/test/java/domain/DealerTest.java
@@ -6,56 +6,21 @@ import domain.card.*;
 import domain.card.cardsGenerator.RandomCardsGenerator;
 import domain.participant.Dealer;
 import domain.participant.Player;
-import domain.participant.Players;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
+import java.util.stream.Stream;
 
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 
 class DealerTest {
 
     private final Deck deck = new Deck(new RandomCardsGenerator());
-
-    @DisplayName("플레이어들의 승패를 결정한다")
-    @Test
-    void test() {
-        //given
-        Cards cards1 = Cards.of(
-                List.of(
-                        new Card(CardNumber.TEN, CardShape.CLOVER),
-                        new Card(CardNumber.THREE, CardShape.CLOVER)
-                )
-        );
-        Cards cards2 = Cards.of(
-                List.of(
-                        new Card(CardNumber.TEN, CardShape.CLOVER),
-                        new Card(CardNumber.TWO, CardShape.CLOVER)
-                )
-        );
-        Cards cards3 = Cards.of(
-                List.of(
-                        new Card(CardNumber.A, CardShape.CLOVER),
-                        new Card(CardNumber.FOUR, CardShape.CLOVER)
-                )
-        );
-        Dealer dealer = new Dealer(cards1, deck);
-        Player player1 = Player.from("플레이어", cards2);
-        Player player2 = Player.from("플레이어2", cards3);
-        Map<Player, GameResult> expected = Map.of(
-                player1, GameResult.LOSE,
-                player2, GameResult.WIN
-        );
-
-        //when
-        Map<Player, GameResult> actual = dealer.getPlayerResult(new Players(List.of(player1, player2)));
-
-        //then
-        assertThat(actual).isEqualTo(expected);
-    }
 
     @DisplayName("플레이어에게 카드를 할당할 수 있다.")
     @Test
@@ -98,5 +63,196 @@ class DealerTest {
             softly.assertThat(count).isEqualTo(2);
             softly.assertThat(dealer.getCardScore()).isEqualTo(17);
         });
+    }
+
+    @DisplayName("승/패/무승부가 명확한 경우, 게임 결과를 올바르게 반환할 수 있다.")
+    @ParameterizedTest
+    @MethodSource("createGameResultCase")
+    void 게임_결과_반환(List<Card> self, List<Card> other, GameResult expected) {
+        // given
+        Cards playerCards = Cards.of(self);
+        Cards dealerCards = Cards.of(other);
+        Player player = Player.from("player", playerCards);
+        Dealer dealer = new Dealer(dealerCards, deck);
+
+        // when
+        GameResult actual = dealer.getResult(player);
+
+        // then
+        assertThat(actual).isSameAs(expected);
+    }
+
+    private static Stream createGameResultCase() {
+        return Stream.of(
+                Arguments.of(
+                        // 21
+                        List.of(
+                                new Card(CardNumber.A, CardShape.CLOVER),
+                                new Card(CardNumber.QUEEN, CardShape.CLOVER)
+                        ),
+                        // 12
+                        List.of(
+                                new Card(CardNumber.TEN, CardShape.CLOVER),
+                                new Card(CardNumber.TWO, CardShape.CLOVER)
+                        ),
+                        "WIN"
+                ),
+                Arguments.of(
+                        // 12
+                        List.of(
+                                new Card(CardNumber.QUEEN, CardShape.CLOVER),
+                                new Card(CardNumber.TWO, CardShape.CLOVER)
+                        ),
+                        // 13
+                        List.of(
+                                new Card(CardNumber.A, CardShape.CLOVER),
+                                new Card(CardNumber.TEN, CardShape.CLOVER),
+                                new Card(CardNumber.TWO, CardShape.CLOVER)
+                        ),
+                        "LOSE"
+                ),
+                Arguments.of(
+                        // 22
+                        List.of(
+                                new Card(CardNumber.TEN, CardShape.CLOVER),
+                                new Card(CardNumber.QUEEN, CardShape.CLOVER),
+                                new Card(CardNumber.TWO, CardShape.CLOVER)
+                        ),
+                        // 22
+                        List.of(
+                                new Card(CardNumber.TEN, CardShape.CLOVER),
+                                new Card(CardNumber.QUEEN, CardShape.CLOVER),
+                                new Card(CardNumber.TWO, CardShape.CLOVER)
+                        ),
+                        "LOSE"
+                ),
+                Arguments.of(
+                        // 20
+                        List.of(
+                                new Card(CardNumber.TEN, CardShape.CLOVER),
+                                new Card(CardNumber.QUEEN, CardShape.CLOVER)
+                        ),
+                        // 20
+                        List.of(
+                                new Card(CardNumber.JACK, CardShape.CLOVER),
+                                new Card(CardNumber.KING, CardShape.CLOVER)
+                        ),
+                        "DRAW"
+                )
+        );
+    }
+
+    @DisplayName("플레이어와 딜러 모두 버스트가 아닌 경우, 합이 더 큰 사람이 승리한다")
+    @Test
+    void 승패_검증() {
+        // given
+        Cards playerCards = Cards.of(List.of(
+                new Card(CardNumber.TEN, CardShape.CLOVER),
+                new Card(CardNumber.THREE, CardShape.CLOVER)
+        ));
+        Cards dealerCards = Cards.of(List.of(
+                new Card(CardNumber.TEN, CardShape.CLOVER),
+                new Card(CardNumber.TWO, CardShape.CLOVER)
+        ));
+        Player player = Player.from("player", playerCards);
+        Dealer dealer = new Dealer(dealerCards, deck);
+
+        // when
+        GameResult actual = dealer.getResult(player);
+
+        //then
+        assertThat(actual).isEqualTo(GameResult.WIN);
+    }
+
+    @DisplayName("플레이어의 카드가 버스트인 경우 패배한다")
+    @Test
+    void 플레이어의_카드가_버스트인_경우_패배한다() {
+        // given
+        Cards playerCards = Cards.of(List.of(
+                new Card(CardNumber.TEN, CardShape.CLOVER),
+                new Card(CardNumber.TEN, CardShape.CLOVER),
+                new Card(CardNumber.TWO, CardShape.CLOVER)
+        ));
+        Cards dealerCards = Cards.of(List.of(
+                new Card(CardNumber.TEN, CardShape.CLOVER),
+                new Card(CardNumber.TWO, CardShape.CLOVER)
+        ));
+        Player player = Player.from("player", playerCards);
+        Dealer dealer = new Dealer(dealerCards, deck);
+
+        // when
+        GameResult actual = dealer.getResult(player);
+
+        //then
+        assertThat(actual).isEqualTo(GameResult.LOSE);
+    }
+
+    @DisplayName("딜러가 버스트인 경우, 버스트가 아닌 플레이어는 승리한다")
+    @Test
+    void 딜러가_버스트인_경우_승리한다() {
+        // given
+        Cards playerCards = Cards.of(List.of(
+                new Card(CardNumber.FIVE, CardShape.CLOVER),
+                new Card(CardNumber.A, CardShape.CLOVER)
+        ));
+        Cards dealerCards = Cards.of(List.of(
+                new Card(CardNumber.TEN, CardShape.CLOVER),
+                new Card(CardNumber.TEN, CardShape.CLOVER),
+                new Card(CardNumber.TWO, CardShape.CLOVER)
+        ));
+        Player player = Player.from("player", playerCards);
+        Dealer dealer = new Dealer(dealerCards, deck);
+
+        // when
+        GameResult actual = dealer.getResult(player);
+
+        //then
+        assertThat(actual).isEqualTo(GameResult.WIN);
+    }
+
+    @DisplayName("모두 버스트가 아니면서 합이 같으면 비긴다")
+    @Test
+    void 두명_모두_버스트가_아니면서_합이_같으면_비긴다() {
+        // given
+        Cards playerCards = Cards.of(List.of(
+                new Card(CardNumber.TEN, CardShape.CLOVER),
+                new Card(CardNumber.TWO, CardShape.CLOVER)
+        ));
+        Cards dealerCards = Cards.of(List.of(
+                new Card(CardNumber.TEN, CardShape.CLOVER),
+                new Card(CardNumber.TWO, CardShape.CLOVER)
+        ));
+        Player player = Player.from("player", playerCards);
+        Dealer dealer = new Dealer(dealerCards, deck);
+
+        // when
+        GameResult actual = dealer.getResult(player);
+
+        //then
+        assertThat(actual).isEqualTo(GameResult.DRAW);
+    }
+
+    @DisplayName("딜러와 플레이어 둘 다 버스트인 경우 플레이어는 패배한다")
+    @Test
+    void 둘_다_버스트인_경우_패배한다() {
+        // given
+        Cards playerCards = Cards.of(List.of(
+                new Card(CardNumber.TEN, CardShape.CLOVER),
+                new Card(CardNumber.TEN, CardShape.CLOVER),
+                new Card(CardNumber.TWO, CardShape.CLOVER)
+        ));
+        Cards dealerCards = Cards.of(List.of(
+                new Card(CardNumber.TEN, CardShape.CLOVER),
+                new Card(CardNumber.TEN, CardShape.CLOVER),
+                new Card(CardNumber.TWO, CardShape.CLOVER)
+        ));
+        Player player = Player.from("player", playerCards);
+        Dealer dealer = new Dealer(dealerCards, deck);
+
+        // when
+        GameResult actual = dealer.getResult(player);
+
+        //then
+        assertThat(actual).isEqualTo(GameResult.LOSE);
     }
 }
